Use multiprocess Queues as a command process.
This means we watch the lichess stream in a separate process and queue
its events. We also queue process finished events, this allows the main
thread to have all "state" and make decisions in a threadsafe manner.

Probably fixes the race conditions. Logs information so that we can
debug if it doesn't