Simplify deferredFirstMove logic
The variable deferredFirstMove is only set to true if the number of half
moves in the game is less than two and it is not the engine's move. So,
when deferredFirstMove is later checked, it is already the engine's
move, so we only need to check if there are less than two half moves to
decide whether to call play_first_book_move() or play_first_move().

Also, since we now explicitly check whether it is the engine's move,
play_first_move() always returns True, so there's no need for it to
return anything.

Also, also, since play_first_book_move() is always followed by
play_first_move(), there's no reason to call play_first_move() within
play_first_book_move(). In fact, it should not be the case that the
function play_first_book_move() can play a non-book move. That makes the
function misnamed.