Merge book config checks into book functions
Instead of checking whether the configuration for the polyglot books are
valid before calling play_first_book_move() or get_book_move(), pass the
configurations to these functions and check there. This also allows for
play_first_book_move() to just call get_book_move() and not check for
itself.