Refactor lichess-bot.py (#270)
* Upgrade pychess library to latest version (1.3.3)

This commit updates lichess-bot to use the latest version of the python
chess library. The main result is a large decrease in the amount of code
in engine_wrapper.py due to so many functions being incorporated into
the chess library.

One annoying difference is that the new library rejects the 'usermove'
feature from Xboard engines, which is rare among chess engine
interfaces.

I hope this can be tested by a variety of engines to make sure it is on
par with the current lichess-bot. This would close #240.

* Remove managed options from engine configs

Some options are automatically managed by the chess engine interface, so
these need to be removed from the engine configurations before creating
the engine.

* Remove parentheses from non-callable members

In the latest version of the python chess library, the result returned
from book lookups has the Move result as a member, not a method, so it
is not callable.

* Remove option that is managed by chess library

This option is also redundant with uci_ponder.

* Use actual latest version of python-chess

* Fix UCI movetime usage

In config file, movetime is in milliseconds, but the library Limit class
expects seconds.

* Handle omitted go_commands

Line 84 prevents self.go_commands from being None

The movetime lines in search_with_ponder() check if there is a movetime
go_command and, if so, convert it from milliseconds to seconds.

* Cosmetic changes

- Delete trailing whitespace
- Standardize spacing around operators
- Delete unneeded parentheses

* Consolidate repeated code

* Fix handling of endgame table paths for Xboard

* Consolidate repeated stats code

* Consolidate repeated stderr logic

* Consolidate repeated move search code

The interface differences between UCI and Xboad protocols have been much
reduced in the latest python-chess library, allowing for much more code
to be shared here.

* Print stats after every move

Since stats are printed after every move, incorporate the printing into
the engine.search() function.

* Simplify deferredFirstMove logic

The variable deferredFirstMove is only set to true if the number of half
moves in the game is less than two and it is not the engine's move. So,
when deferredFirstMove is later checked, it is already the engine's
move, so we only need to check if there are less than two half moves to
decide whether to call play_first_book_move() or play_first_move().

Also, since we now explicitly check whether it is the engine's move,
play_first_move() always returns True, so there's no need for it to
return anything.

Also, also, since play_first_book_move() is always followed by
play_first_move(), there's no reason to call play_first_move() within
play_first_book_move(). In fact, it should not be the case that the
function play_first_book_move() can play a non-book move. That makes the
function misnamed.

* Merge book config checks into book functions

Instead of checking whether the configuration for the polyglot books are
valid before calling play_first_book_move() or get_book_move(), pass the
configurations to these functions and check there. This also allows for
play_first_book_move() to just call get_book_move() and not check for
itself.

* Consolidate first move functions

* Simplify mid-game move-picking logic

For the first two moves of the game, the logic goes:
If there's a book move, use it. Otherwise, search for a move.

For all other moves:
1. Wait to stop pondering.
2. If there's a book move, use it.
3. If the pondering move was correct, use it.
4. Otherwise, search.

* Create function for mid-game moves

* Create function to start pondering thread

Also, both play_move() and start_pondering() need the start_time
variable, so move that outside play_move().

* Delete unneeded line

* Delete redundant else case

* Unify first move and rest of game loops

* Combine ping code for black and white

* Use guard-clause style to reduce indentation

* Add blank line

* UCI bots can ponder after first move

* Move side effect from functions to main loop

* Create function for getting ponder results

* Create function for fake thinking

* Move ponder_thread_func() to start_pondering()

* Add missing parameter

* Further code cleanup

1. Remove unused parameters from choose_first_move() and choose_move()
2. Use guard-clause style and early returns to reduce indentation
3. Order move parameters consistently

* Move book and pondering functions into main loop

Moving these functions into the main loop reduces the amount of
duplicated code in the move choosing functions and makes the logic flow
clearer:

1. Is there a book move? If yes, use it. Else
2. Is it the first move of the game? If yes, choose the first move. Else
3. Is there a result from pondering? If yes, use it. Else
4. Choose the next move by actually thinking about it.

Separating the book and pondering functions also decreased the number of
parameters being passed to each function.

* Time used during fake thinking is accounted for

In order for the engine to accurately know how much time it has, it
needs to know how much time was spent during fake thinking.

Also, change the name of the variable in choose_move() to more
accurately reflect its purpose.

* Add blank line

* Group exception handlers in main game loop

* Get book configurations in get_book_move()

Further reduce the number of variables in play_game() and reduce the
number of arguments in get_book_move()