Disable pondering after first move
Because the first move on lichess uses a different clock than the rest
of the game, engines cannot ponder after the first move because the
python-chess library sets the clock for pondering to be the same as the
move. After black's first move, it will think it has 10 seconds to
ponder no matter how short the time control and no matter how little
time white takes for its second move.