Add small move delay to avoid rate limiting
This might be an edge case, but my bot plays moves instantly, and when it plays multiple players in hyperbullet, keyboard users or other bots that move instantly, it would sometimes run into an 429 error. The problem is that after running into this error, the bot does not resume play on that game, and will consequently lose on time. My solution might not be sophisticated but it gets the job done, and the delay is negligible. 

After adding the delay, my bot was able to exploit this bug against a few other instant-moving bots, to make them lose on time. Here are some examples:
https://lichess.org/tXHK71bx/black
https://lichess.org/4qBo7GZ0/black
https://lichess.org/MOBAl4vY

In the first game, 26 moves were played within 0.3 seconds, at which point the opponent bot stopped making moves, presumably due to the same issue as I had. This shows that it's a real issue affecting real bots.

I believe that the behavior stems from is_final() in lichess.py, which considers 429 errors to be final, and will therefore not retry after receiving them. But I believe this is the intended behavior, so this shouldn't be changed? Perhaps another solution would be to only delay moves if the previous move was played very recently. This way, it would not affect slow games at all.