Allow XBoard bots of offer/accept draws and resign
These changes allow XBoard engines to resign games (by printing "resign"
to stdout) and to offer/accept draws (by printing "offer draw" to
stdout). There are no similar functions for UCI engines.

Two problems so far:

1. The python-chess library will only pick up on "offer draw" if the bot
prints it while thinking about a move. Otherwise, it is ignored. So, if
the bot responds immediatlely to the "draw" command on line 262 before
one of the choose_move functions can be called, it will be ignored.

2. When accepting a draw, the lichess API returns with

    https://lichess.org:443 "POST /api/bot/game/za8D4j9z/move/c1f4?offeringDraw=1 HTTP/1.1" 400 43
    Giving up api_post(...) after 1 tries (requests.exceptions.HTTPError: 400 Client Error: Bad Request for url: https://lichess.org/api/bot/game/za8D4j9z/move/c1f4?offeringDraw=1)

The draw acceptance is properly accepted by lichess and the game
actually ends with a draw. I'm not sure if the error message is
important.