Moved config to test_bot
The config was moves to test_bot, so it will be easier to implement new test in the future. To make a new test, one would have to make a new config and a new download function for the engine so they don't have to change the run_bot function. If someone decides to make a new test they should use an engine that is consistently able to beat the AI (currently at level 4 but can be changed).