Fix Intro message (#410)
The intro message never gets sent in a game, but the message at the end of the game gets sent. This commit fixes that by placing the line to send the intro message (`conversation.send_message("player", hello)`) at the start of the game and making sure it sends the message only once every game.