Start engines outside of engine folder
The idea here is to not require users to place their engine executables
into the engines/ folder in lichess-bot. The engine: dir configuration
can take an absolute path and use that for the working directory if the
engine needs to read or write files.

As it is now, the engine and reading/writing files all have to be in the
same place. Next work should be adding a working directory configuration
so that the executable and the data can be in different places.