Explicitly set a working directory
There are now two directories in the engine: section of the config
YAML file. The configuration `dir` specifies the directory containing
the bot executable. The configuratoin `working_dir`, specifies the
base directory where files are read and written. If `working_dir` is
missing or blank, it defaults to `dir`.

Add a sanity check for the `working_dir` that mirrors the `dir`
check.

Update README documentation