Separate working directory (#416)
This commit lets the user specify which directory the bot should run from. This is useful where the bot needs to read files or writes files and the user does not want them in the lichess directory.

If the new working directory configuration is blank or missing, the current working directory is used, which is the current behavior. This change should be backward compatible with older config files.

This implements the request in issue #232.

===

* Start engines outside of engine folder

The idea here is to not require users to place their engine executables
into the engines/ folder in lichess-bot. The engine: dir configuration
can take an absolute path and use that for the working directory if the
engine needs to read or write files.

As it is now, the engine and reading/writing files all have to be in the
same place. Next work should be adding a working directory configuration
so that the executable and the data can be in different places.

* Explicitly set a working directory

There are now two directories in the engine: section of the config
YAML file. The configuration `dir` specifies the directory containing
the bot executable. The configuratoin `working_dir`, specifies the
base directory where files are read and written. If `working_dir` is
missing or blank, it defaults to `dir`.

Add a sanity check for the `working_dir` that mirrors the `dir`
check.

Update README documentation

* Use correct default working directory

* Rename kwargs to popen_args to indicate function

Eventually, the cwd option will make its way to a popen command.

* Correct documentation to match code

* Correct comment in default config file

* Delete unused EngineWrapper.__init__() arguments

* Revert lines that didn't need changing