Check that correspondence_queue is empty not False
Since Python Queues are meant for multithreaded contexts, it does not
act like other Python collections. In this case, an empty queue is not
considered False. So, the loop on line 192 never exits because the
condition on line 196 is always true after a corresondence ping. This
change makes the check for empty explicit.