Suppress API error messages by default / Keep connection after game timeout (#429)
* Get last message after game clock expires

If a bot runs out of time, this change will make sure that the last message
from lichess is received. Prior to this change, the user would only see
an api_post() error and the game would end.

* Set logging level of backoff to DEBUG

This commit causes "Giving up api_post(...)" error messages to display
only when lichess-bot is run with -v. Otherwise, the backoff messages
are hidden.