Record games played by bot to local files (#444)
* Record games to local files wtih bot thinking

Not yet able to recover from losing connections or restarting in the
middle of a game.

* Add header information for chess variants

* Make sure every move has a comment

In case an engine does not output mid-search evaluations on all moves,
this change ensures that the commentary can be lined up with the
game moves when printing the pgn file.

* Better name for list variable

* Proper time control formatting

* Handle bot runs out of time

If the bot runs out of time, it will have an extra move in its
commentary that is not in the game record. This would cause record
writing to fail when the first move of the game with commentary could
not be found.

* Game record errors don't stop other commands

* Make aligning moves with comments more robust

* Add termination header to game record

* Document PGN feature

* Delete unneeded code

* Add comments to game record writer sections

* Write game record for engines that don't output thinking

* Remove question marks from game record file names

Question marks are not allowed in Windows file names.

* Write game records during tests

* Handle when not all moves have comments

* Spacing for legibility

* Handle when commentary ends before last move

* Style: single-quotes to double-quotes

* Remove all characters invalid in Windows file names

* Warn user if game record writing fails

Letting an exception propagate will restart the play_game() function,
which is not the right action. Just tell the user that the game record
writing failed and a partial reason.

* Make error message more specific

* Properly escape backslash

* Use python-chess methods to add comments to PGN

* Simpler check for pre-existing game record file

* Add tests to check that PGN files are written

* Simplify code

Delete unneeded variables. Remove unneeded try-except blocks.

* Don't add Termination header if game is not finished

* Line up arguments

* Add note about recorded scores

* Better documentation of PGN record scores