Fix duplicate variation lines written to PGN file (#455)
Previously, if a bot had no commentary for any moves, spurious variation lines
that were identical to the main line would be written to the first move in
the PGN record. This line would be copied every time lichess-bot reconnected
to a game in progress and the bot didn't play a move. This can happen when
a bot checks in on a correspondence game and no opponent moves have
been played.

The error in the code is that the index found in the changed section of
code would return 0 since the result of calling all() on an empty set is true.

This commit fixes the problem by aligning to the end of the game record
if there is no commentary.