Remove newlines from between joined comments
Lichess sends moves with multiple comments, these comments are
separately delimited: {First} {Second}. The default PGN parser in
python-chess joins the comments with a newline. This commit creates
a slightly modified parser that joins these comments with a space.

There is a different change that could be done with a regex one-liner, but
that relies on both comments being contained in the same line.