Fix error when db move is chosen after engine move
If a database or book move is chosen in the middle of a game and there are
engine-chosen moves before that move, then the engine's comments
will no longer be aligned with the moves on the board. This commit
preserves that alignment by inserting None as the comment to non-engine
moves after the first engine move.

The code was further cleaned up by creating new EngineWrapper methods.