Add go commands to all moves (#468)
The go commands from the user's config.yml file are not used during the first move of a game nor during correspondence games. This PR fixes that.

Other features, replace integer division (//) with normal division (/) so that move times are printed with sub-second accuracy.

Closes #466

Commit history:
* Use go commands from config file in all moves

* Use correct field name

* Delete trailing whitespace

* Return new time limit value

* Use correct field name

* Correct usage of movetime

The correct comparison is when both movetime and time_limit.time
are in seconds.