Combine challenge accept and decline
When checking if a challenge is supported in a bot's configuration,
return a tuple of True/False for whether the challenge is accepted and
a string/None for the reason it was rejected.

Also, support methods should take the configuration as an argument
instead of parsing it in the calling body.