Handle non-move game updates
When there are changes to the game that are not moves (draw offers
and move takeback requests), lichess sends a gameState message with
the same moves as before the request. This can sometimes cause lichess
to signal to the bot to consider the prior board position that it already
played. This update fixes this issue and adds a place for a bot to accept
its opponent's move takeback requests if configured to do so (not
implemented yet).

Fixes #502