Use separate method for online chess DB access
Since the api_get() method uses backup to retry communications, any
problems with communicating with the online chess databases can
result in retrying for up to a minute. This would waste time that bots
could use to pick their own moves.

This commit replaces the Lichess.api_get() method with another that
does not have backoff, so it does not retry requests.