More fine-grained delay control
Whether a bot is challenged again depends not just on whether that bot
declines a challenge, but also on the type of game that was declined.
Lichess-bot will delay challenging a bot if that bot has previously refused
a game with the same chess variant, time control speed, and rated option.