Replace long if-elif-else blocks with function
The new range_scan() function replaces if-elif-else blocks that assign
scores to ranges of integers.

The first argument is a list of 2-tuples sorted by the first element: the
first element defines the inclusive right edge of the range and the second
element is the return value should the queried position be inside that range.

The second argument is the value to be returned if the queried position does
not fit in any ranges defined by the first argument.

The third argument is the queried position.

I will admit the function results in the code looking rather LISP-ish, especially
near the bottom of get_gaviota().