Suggest moves from EGTBs to the engine (#588)
Suggests moves from EGTB to the engine. The moves that are suggested are all the moves that could have been chosen with move_quality="good". This will fix the problem of EGTBs playing dumb moves (e.g. 8/8/8/8/1p6/1K1R4/8/2k5 w - - 0 1), where the engine instead of mating will first pick up material. Getting the move from lichess will fix this problem for the most part because it also has DTM, but it doesn't have DTM for 6 or 7 men positions.

This can also be done with lichess_cloud_analysis but I don't know if this is useful.

Also had to move some code to fix the flake8 errors that the initial addition created.

* Don't allow user to use suggest with XBoard engines

* Simplify code and fix bug